docs: refactor nextjs server-side auth to match proxy
