+++ examples/auth/nextjs-full/proxy.ts
@@ -1,7 +1,7 @@
 import { type NextRequest } from "next/server"
-import { updateSession } from "@/utils/supabase/middleware"
+import { updateSession } from "@/lib/supabase/proxy"
 
-export async function middleware(request: NextRequest) {
+export async function proxy(request: NextRequest) {
   return await updateSession(request)
 }
